Simulating Neutral Landscape Models

NLMR is an R package for simulating neutral landscape models (NLM).

Details

This package contains vignettes that introduce NLM and basic usage of the NLMR package. The vignettes in this package are listed below.

NLM Concepts + Terminology

Background information on the concepts and terminology that underpin the simulation of neutral landscape models.

NLMR Bestiary

Bestiary of the available neutral landscape models.

Quickstart Guide

Short walk-through of the NLMR package and how to handle the simulations.

See also

Useful links:

https://marcosci.github.io/NLMR/

Report bugs at https://github.com/marcosci/NLMR/issues/