distancegradientNLM

Create a distance gradient neutral landscape model with values ranging 0-1.

distancegradientNLM(nCol, nRow, origin, rescale = TRUE)

Arguments

nCol
[numerical(1)] Number of columns for the raster.

nRow
[numerical(1)] Number of rows for the raster.

origin
[numerical(4)] Edge coordinates of the origin of the distance measurement.

rescale
[logical(1)] If TRUE (default), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated.

Value

RasterLayer with distance measurements starting in a predefined origin.

Details

origin is a umeric vector of xmin, xmax, ymin, ymax for a square inside the raster from which the distance is measured. xmax and ymax must be smaller than nCol and nRow.

Examples

distancegradientNLM(nCol = 100, nRow = 100, origin = c(20, 30, 10, 15))

#> class       : RasterLayer 
#> dimensions  : 10, 10, 100  (nrow, ncol, ncell)
#> resolution  : 10, 10  (x, y)
#> extent      : 0, 100, 0, 100  (xmin, xmax, ymin, ymax)
#> coord. ref. : NA 
#> data source : in memory
#> names       : layer 
#> values      : 0, 1  (min, max)
#>