metric_area

Report area for each of the constituent values making up a matrix

metric_area(x, poi = NULL)

Arguments

x
[Raster* object]

poi
[numerical] Vector of numeric values indicating classes to be reported.

Value

List of tibbles

Details

Function reports the number of cells and the proportion of total cells made up by each unique class value

Examples

x <- nlm_random(100, 100)
y <- c(0.5, 0.25, 0.25)
z <- util_classify(x, y)

metric_area(z)

#> $Total_Area
#> # A tibble: 3 x 2
#>   count class
#>   <dbl> <int>
#> 1  4999     0
#> 2  2500     1
#> 3  2501     2
#> 
#> $Proportion_Area
#> # A tibble: 3 x 2
#>   count class
#>   <dbl> <int>
#> 1 0.500     0
#> 2 0.250     1
#> 3 0.250     2
#>