nlm_planargradient

Simulates a planar gradient neutral landscape model.

nlm_planargradient(ncol, nrow, resolution = 1, direction = NA,
  rescale = TRUE)

Arguments

ncol
[numerical(1)]
Number of columns forming the raster.

nrow
[numerical(1)]
Number of rows forming the raster.

resolution
[numerical(1)]
Resolution of the raster.

direction
[numerical(1)]
Direction of the gradient in degrees, if unspecified the direction is randomly determined.

rescale
[logical(1)]
If TRUE (default), the values are rescaled between 0-1.

Value

RasterLayer

Details

Simulates a linear gradient sloping in a specified or random direction.

References

Palmer, M.W. (1992) The coexistence of species in fractal landscapes. The American Naturalist, 139, 375 - 397.

See also

Examples

# simulate planar gradient
planar_gradient <- nlm_planargradient(ncol = 200, nrow = 200)
# NOT RUN {
# visualize the NLM
rasterVis::levelplot(planar_gradient, margin = FALSE, par.settings = rasterVis::viridisTheme())
# }