nlm_polylands

Simulate the NLM introduced in Gaucherel 2008.

nlm_polylands(nCol, nRow, resolution = 1, option = 1, germs, g, R,
  patch_classes, rescale = TRUE)

Arguments

nCol
[numerical(1)] Number of columns for the raster.

nRow
[numerical(1)] Number of rows for the raster.

resolution
[numerical(1)] Resolution of the raster.

option
[numerical(1)] If 1 (default), the Tessellation method is used to simulate the NLM. If 2, the Gibbs algorithm method is used to simulate the NLM.

germs
[numerical(1)] Intensity parameter (non-negative integer).

g
[numerical(1)] Interaction parameter (a value between 0 and 1, inclusive - only used when option = 2).

R
[numerical(1)] Interaction radius (non-negative integer).

patch_classes
[numerical(1)] Number of classes for germs.

rescale
[logical(1)] If TRUE (default), the values are rescaled between 0-1. Otherwise, the distance in raster units is calculated.

Value

RasterLayer

Details

The function offers 2 of the 3 NLM described in Gaucheral. The first one (option = 1) is a tessellation method. It generates a random point pattern (the germs) with an independent distribution and uses the voronoi tessellation to simulate the patchy landscapes. The second one (option = 2) is the Gibbs algorithm method. The method works in principal like the tessellation method, but instead of a random point pattern one fits a simulated realization of the Strauss process. The Strauss process starts with a given number of points and uses a minimization approach to fit a point pattern with a given interaction parameter (0 - hardcore process; 1 - poission process) and interaction radius (distance of points/germs being apart).

References

Gaucherel, C. (2008) Neutral models for polygonal landscapes with linear networks. Ecological Modelling, 219, 39 - 48.

Examples

nlm_polylands(nCol = 50, nRow = 50, germs = 20)

#> class       : RasterLayer 
#> dimensions  : 50, 50, 2500  (nrow, ncol, ncell)
#> resolution  : 1, 1  (x, y)
#> extent      : 0, 50, 0, 50  (xmin, xmax, ymin, ymax)
#> coord. ref. : NA 
#> data source : in memory
#> names       : levels 
#> values      : 0, 1  (min, max)
#>