nlm_randomrectangularcluster

Create a random rectangular cluster neutral landscape model with values ranging 0-1.

nlm_randomrectangularcluster(ncol, nrow, resolution = 1, minL, maxL,
  rescale = TRUE)

Arguments

ncol
[numerical(1)] Number of columns for the raster.

nrow
[numerical(1)] Number of rows for the raster.

resolution
[numerical(1)] Resolution of the raster.

minL
[numerical(1)] The minimum possible width and height for each random rectangular cluster.

maxL
[numerical(1)] The maximum possible width and height for each random rectangular cluster.

rescale
[logical(1)] If TRUE (default), the values are rescaled between 0-1.

Value

RasterLayer with random values ranging from 0-1.

Examples

# simulate random rectangular cluster
randomrectangular_cluster <- nlm_randomrectangularcluster(ncol = 30,
                                                          nrow = 30,
                                                          minL = 5,
                                                          maxL = 10)
# NOT RUN {
# visualize the NLM
util_plot(randomrectangular_cluster)
# }