nlm_randomcluster

Simulates the random cluster nearest-neighbour algorithm.

nlm_randomcluster(nCol, nRow, resolution = 1, neighbourhood = 8, p,
  rescale = TRUE)

Arguments

nCol
[integer(1)] Number of columns in the raster.

nRow
[integer(1)] Number of rows in the raster.

resolution
[numerical(1)] Resolution of the raster.

neighbourhood
[numerical(1)] Clusters are defined using a set of neighbourhood structures, 4 (Rook's case), 8 (Queen's case).

p
[numerical(1)] The p value controls the proportion of elements randomly selected to form clusters.

rescale
[logical(1)] If TRUE (default), the values are rescaled between 0-1.

Value

Raster with random values ranging from 0-1.

Details

The algorithm (Etherington et al. 2014) is an adoption of the MRC algorithm by Saura & Martínez-Millán (2000) The algorithm simulates a percolation map, which defines random clusters by running a connected labellign algorithm which detects clusters and gives them a unique ID. It controls the size and directional bias of the cluster with the proportion of the matrix that is within a cluster and with specifying a specific neighbourhood rule. Each cluster is than given a random value and non-cluster cells are assigned values by performing a nearest neighbour interpolation.

References

Saura, S. & Martínez-Millán, J. (2000) Landscape patterns simulation with a modified random clusters method. Landscape Ecology, 15, 661–678. Etherington TR, Holland EP, O’Sullivan D. 2015. NLMpy: A python software package for the creation of neutral landscape models within a general numerical framework. Methods in Ecology and Evolution 6:164–168.

Examples

nlm_randomcluster(nCol = 10, nRow = 10, neighbourhood = 4, p = 0.4)

#> class       : RasterLayer 
#> dimensions  : 10, 10, 100  (nrow, ncol, ncell)
#> resolution  : 1, 1  (x, y)
#> extent      : 0, 10, 0, 10  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0 
#> data source : in memory
#> names       : layer 
#> values      : 0, 1  (min, max)
#>